+++ lib/just_bash/commands/paste.ex
@@ -3,6 +3,7 @@ defmodule JustBash.Commands.Paste do
   @behaviour JustBash.Commands.Command
 
   alias JustBash.Commands.Command
+  alias JustBash.Commands.StdinOperand
   alias JustBash.FS
 
   @impl true
@@ -43,7 +44,11 @@ defmodule JustBash.Commands.Paste do
   end
 
   defp parse_args(args) do
-    parse_args(args, %{delimiter: "\t", serial: false, files: []})
+    {option_args, extra} = StdinOperand.split_end_of_options(args)
+
+    with {:ok, opts} <- parse_args(option_args, %{delimiter: "\t", serial: false, files: []}) do
+      {:ok, %{opts | files: opts.files ++ extra}}
+    end
   end
 
   defp parse_args([], opts), do: {:ok, opts}