+++ lib/just_bash/commands/shasum.ex
@@ -45,7 +45,9 @@ defmodule JustBash.Commands.Shasum do
   defp defaults_to_stdin(files), do: files
 
   defp parse_args(args) do
-    parse_args(args, %{algorithm: "1", check: false}, [])
+    {option_args, extra} = StdinOperand.split_end_of_options(args)
+    {opts, files} = parse_args(option_args, %{algorithm: "1", check: false}, [])
+    {opts, files ++ extra}
   end
 
   defp parse_args([], opts, files), do: {opts, Enum.reverse(files)}